Import -> Include ; Taxonomy base
Replace 'import' with 'include' because Java version treats 'import' as reserved word.
Added 'Taxonomy' as base for LogicalEnumeratedBase.

NOTE also: previously the default grammar used '(' and ')' as delimiters for LogicalEnumeratedSet. This results in an ambiguity when generating the grammar. It was replaced with '{' and '}' to get it to build properly. This ambiguity  may have been addressed? (see https://github.com/TypeFox/xtext2langium/issues/9)